Adding permission constraint while using APIS

+++ src/blog/api/views.py
@@ -1,12 +1,14 @@
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
 from account.models import Account
 from blog.models import BlogPost
 from blog.api.serializers import BlogPostSerializer
 
 
 @api_view(['GET', ])
+@permission_classes((IsAuthenticated,))
 def api_detail_blog_view(request, slug):
 
     try:
@@ -20,6 +22,7 @@ def api_detail_blog_view(request, slug):
 
 
 @api_view(['PUT', ])
+@permission_classes((IsAuthenticated,))
 def api_update_blog_view(request, slug):
 
     try:
@@ -27,6 +30,10 @@ def api_update_blog_view(request, slug):
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
+
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response':"You don't have permission to edit this post." })
     if request.method == 'PUT':
         serializer = BlogPostSerializer(blog_post, data=request.data)
         data = {}
@@ -38,12 +45,17 @@ def api_update_blog_view(request, slug):
 
 
 @api_view(['DELETE', ])
+@permission_classes((IsAuthenticated,))
 def api_delete_blog_view(request, slug):
 
     try:
         blog_post = BlogPost.objects.get(slug=slug)
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
+
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response':"You don't have permission to delete this post." })
 
     if request.method == 'DELETE':
         operation = blog_post.delete()
@@ -56,9 +68,10 @@ def api_delete_blog_view(request, slug):
 
 
 @api_view(['POST', ])
+@permission_classes((IsAuthenticated,))
 def api_create_blog_view(request):
 
-    author = Account.objects.get(pk=1)
+    author = request.user
 
     blog_post = BlogPost(author=author)
